finish forgejo
This tests a technique of generating a vendor bundle directory
which is downloaded and installed as a second distribution file.
Every upgrade will require a new version and it will require
us to host it.

A new script will come to make this process easier.  This just
serves as a proof of concept.